Move tests to their own folder
